refactor: 컨벤션에 맞추어 줄바꿈
