+++ backend/src/main/java/com/staccato/image/domain/ImageExtension.java
@@ -1,4 +1,4 @@
-package com.staccato.image.service;
+package com.staccato.image.domain;
 
 import java.util.Arrays;
